python_backend/requirements.txt
fastapi
uvicorn
sentence-transformers
torch
pydantic
supabase
python-dotenv
requests
llama-cpp-python
huggingface_hub
